Move contracts' tests to subfolder
